Defined grads of vspace methods using defjvp and defvjp_argnums, allowing the rest of the grad-defining wrappers to be moved to extend.py
